% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Massign.R
\name{\%*>\%}
\alias{\%*>\%}
\title{Construct a matrix from string and matrix multiply it}
\usage{
string \%*>\% second_arg
}
\arguments{
\item{string}{a matrix in character form to be converted to a numeric matrix.}

\item{second_arg}{either a matrix, or a character to be converted to a matrix
by the  \code{\%<-\%} operator.}
}
\description{
Interpret the character form of a matrix, and multiply it.
}
\details{
This function can be used quickly perform tests with matrix multiplication by
giving the matrices in character form as interpreted by the \code{\%<-\%}
operator. It also fits well in a pipe-operator syntax.
}
\examples{

# Use to quickly test the matrix multiplication of two matrices interpreted from strings.
"1, 2
 3, 4" \%*>\%
 " 0, 1
   1, 0"

# Use when the second argument is a matrix.
"1, 2, pi \\\\ 3, 4, 1 \\\\ 3, 2, 1" \%*>\% diag(c(1, 2, 3))


}
