% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multiplipe.R
\name{Multiplipe}
\alias{Multiplipe}
\alias{\%*>\%}
\title{Quickly test matrix multiplication of two matrices interpreted from strings.}
\usage{
matrix1 \%*>\% matrix2
}
\arguments{
\item{matrix1}{a matrix or Massign character matrix that premultiplies}

\item{matrix2}{a matrix or Massign character matrix that postmultiplies}
}
\description{
Building on Massign's core functionality, the Multiplipe operator `%*>%`
allows for quick prototyping of matrix multiplications.
}
\examples{
# Basic usage
"1, 2
 3, 4" \%*>\%
 " 0, 1
   1, 0"

# Second argument can be a matrix:
"1, 2, pi \\\\ 3, 4, 1 \\\\ 3, 2, 1" \%*>\% diag(c(1, 2, 3))

# Or the first, for that matter:
diag(c(1, 2, 3)) \%*>\% "1, 2, pi \\\\ 3, 4, 1 \\\\ 3, 2, 1"


}
\seealso{
\code{\link[base]{matrix}}, \code{\link{Massign}}
}
