% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Massign.R
\name{Massign}
\alias{Massign}
\alias{\%<-\%}
\alias{\%->\%}
\title{Construct a matrix from a (formatted) string}
\usage{
var \%<-\% value

value \%->\% var
}
\arguments{
\item{var}{the variable to which the matrix will be assigned. Can be an element
of a list.}

\item{value}{a matrix in character form to be converted to a numeric
matrix.}
}
\description{
Constructing matrices for quick prototyping can be very annoying in \code{R},
requiring the user to think about how to fill the matrix with values
using the \code{matrix(data, nrow, ncol, byrow)} function. The \code{\%<-\%}
operator solves that issue by allowing the user to construct string matrices
that look like actual matrices.
}
\examples{
M \%<-\% "   1,  0.2, -0.3,  0.4
         0.2,    1,  0.6, -0.4
        -0.3,  0.6,    1,  0.4
         0.4, -0.4,  0.4,    1"

}
\seealso{
\code{\link[base]{matrix}}
}
